refactor domain product detail route
